move some dtype stuff to cpp_type_aux.py
